+++ EATS/src/main/java/com/userexprior/EatsApplication.java
@@ -3,11 +3,13 @@ package com.userexprior;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
 @ComponentScan(basePackages = "com.userexprior")
 @EnableScheduling
+@EnableJpaRepositories(basePackages = "com.userexprior.repository")
 public class EatsApplication {
 
 	public static void main(String[] args) {
