+++ EATS/src/main/java/com/userexprior/controller/EmployeeController.java
@@ -3,13 +3,11 @@ package com.userexprior.controller;
 import java.util.List;
 import java.util.Map;
 
+import com.userexprior.service.EmployeeAdaptor;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.client.RestTemplate;
 
-import com.userexprior.adaptor.EmployeeAdaptor;
 import com.userexprior.service.EmployeeService;
 
 
@@ -21,13 +19,14 @@ public class EmployeeController {
 	 * Injected automatically by Spring's dependency injection.
 	 */
 	@Autowired
-	EmployeeService employeeService;
+	private EmployeeService employeeService;
 
 	/**
 	 * Adapter for transforming employee data into UI-friendly format.
 	 * Converts raw employee data into structured response objects.
 	 */
-	EmployeeAdaptor employeeAdaptor = new EmployeeAdaptor();
+	@Autowired
+	private EmployeeAdaptor employeeAdaptor;
 
 	/**
 	 * Retrieves all employee data and their activities.
