+++ EATS/src/main/resources/application.properties
@@ -3,7 +3,7 @@ spring.datasource.username = root
 spring.datasource.password = root
 #spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver 
  
-
+server.port=8082
 ## Hibernate Properties added
 # The SQL dialect makes Hibernate generate better SQL for the chosen database
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL8Dialect
